ArrayExample.py
First for loop shows what hardcoding is and how limited it is. The second for loop shows a more flexible for loop using the range of aryLength, it uses i to loop through aryLength. Also when printed out, i will print what position each fruit is in, for example, cherries would be position 0.